@@ -101,10 +101,7 @@ class TestDevicesList:
         r = c.run(
             "--server", "https://mender-api-gateway", "--skip-verify", "devices", "list", "--raw"
         )
-        pattern = r'(\[\{.*?\}\](?=\s|$))' # Regex for locating json in stdout
-        matches = re.findall(pattern, r.stdout, re.DOTALL)
-        raw_string = max(matches, key=len)
-        raw = json.loads(raw_string, strict=False)
+        raw = json.loads(r.stdout, strict=False)
         assert r.returncode == 0, r.stderr
         assert raw[0]["id"] == device_id
         assert raw[0]["auth_sets"][0]["id"] == authset_id
